Update build_loop.yml
This entire page was copied from the public repository, then I cleared my fucked up on out and pasted that into this. then I added on line 279 the customization code. and trying to commit changes and I get that stupid error.. I have un saved changes..   then it doesn't work and in actions it fails to run.